[PP][OPS-12139] Giving P800Reference and UserEnteredP800Reference dedicated types (#170)

+++ app/controllers/WeAreVerifyingYourBankAccountController.scala
@@ -178,7 +178,7 @@ class WeAreVerifyingYourBankAccountController @Inject() (
         ),
         paymentData  = Some(PaymentData(
           paymentAmount = Some(journey.getAmount.inPounds), //TODO: according to analysis this comes form repayment status API, but we don't call it.
-          paymentNumber = Some(journey.getP800Reference.sanitiseReference.value.toInt) //TODO: according to analysis this comes form repayment status API, but we don't call it.
+          paymentNumber = Some(journey.getP800Reference.sanitiseReference.value) //TODO: according to analysis this comes form repayment status API, but we don't call it.
         )),
         employerData = None, //TODO: according to the analysis: confirm this is not needed
